third_party/blink/web_tests/flag-specific/highdpi/virtual/scroll-unification/fast/forms/select-popup/popup-menu-mouse-operations-expected.txt

PASS picker.selectElement_.value is "1"
PASS menuElement.value is "bar"
PASS picker.selectElement_.value is "0"
PASS menuElement.value is "bar"
FAIL picker.selectElement_.value should be 1. Was 2.
PASS menuElement.value is "bar"
FAIL picker.selectElement_.value should be 1. Was 2.
PASS menuElement.value is "bar"
FAIL Unexpected error: Uncaught TypeError: Cannot read properties of null (reading 'frameElement')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

foo
bar
baz
 
option1
option2
option3
option4
option5
option6
option7
option8
option9
option10
option11
option12
option13
option14
option15
option16
option17
option18
option19
option20
option21
